// react_code/NHL_App_TS/types/SeriesStatus.tsx
export default class SeriesStatus {
    round: number = 0;
    seriesAbbrev: string = '';
    seriesTitle: string = '';
    seriesLetter: string = '';
    neededToWin: number = 0;
    topSeedTeamAbbrev: string = '';
    topSeedWins: number = 0;
    bottomSeedTeamAbbrev: string = '';
    bottomSeedWins: number = 0;
    gameNumberOfSeries: number = 0;

}